requirements.txt
deep-ocr>=0.1.0
transformers==4.46.3
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
numpy>=1.24.0
tqdm>=4.65.0
PyYAML>=6.0
safetensors>=0.3.0
accelerate>=0.20.0
flash-attn>=2.0.0
